Fix comment
