+++ Korenko HM 24/Korenko HM 24/CellForMainTableView/DailyWeatherCell/DailyCell/DailyWeatherForTableCell.swift
@@ -34,8 +34,8 @@ class DailyWeatherForTableCell: UITableViewCell {
         DispatchQueue.main.async { [weak self] in
             guard let self = self,
                   let weather = weatherData.weather.first else { return }
-            self.timeLabel.text = confertUnixTypeToNormal(time: Double(weatherData.dt), typeTime: "MMM d")
-            self.iconView.image = getIconImage(iconId: weather.icon)
+            self.timeLabel.text = self.confertUnixTypeToNormal(time: Double(weatherData.dt), typeTime: "MMM d")
+            self.iconView.image = self.getIconImage(iconId: weather.icon)
             self.tempLabel.text = "\(Int(weatherData.temp.min)) - \(Int(weatherData.temp.max)) C"
         }
     }
