+++ Korenko HM 24/Korenko HM 24/ViewController /ListViewController.swift
@@ -46,7 +46,7 @@ extension ListViewController: UITableViewDataSource, UITableViewDelegate {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: TableViewCellForListReqest.key) as? TableViewCellForListReqest else { return UITableViewCell() }
         let item = items[indexPath.row]
         guard let weather = item.weather else { return UITableViewCell() }
-        let time = getTime(time: item.time)
+        let time = Int(item.time.timeIntervalSince1970)
         cell.informationLabel.text = "lat = \(item.lat), lot = \(item.lon), time = \(time), temp = \(weather.temp), feelsLike = \(weather.feelsLike)"
         cell.iconView.image = getIconImage(iconId: weather.icon)
         return cell
